Implemented most of block handling
* Added support for 0x80 bytes blocks used to refers to parts of images in the wan file format. So now most images are aligned properly.
- Some images still don't align properly in "effect.bin"
* Made frame parts blend together properly when rendering a frame.
* Attempted to make the code compatible with Qt 5.12, so it can be built on linux and hopefully macos.
* Added more details to the "unknown items" tab in the editor, like the size and header of the particular item.
* Added some method variants to obtain redered frames and images as pixmaps instead of QImage so that can be optimized one day.
* Prototyped block references handling in the frame editor tab. It would probably be better to use a single combobox "image source" or something like that. With a toggle between references and an image. So that sprites that don't use image references at all could be seamlessly handled without the user neededing to know about that.

+++ src/data/sprite/models/images_list_model.cpp
@@ -62,7 +62,7 @@ QVariant ImageListModel::data(const QModelIndex &index, int role) const
         case eColumns::Depth: //depth
         {
             if( role == Qt::DisplayRole )
-                res.setValue(QString("%1bpp").arg(img->getImageOriginalDepth()));
+                res.setValue(QString("%1bpp").arg(static_cast<int>(img->getImageOriginalDepth())));
             break;
         }
         case eColumns::Resolution: //resolution
@@ -172,7 +172,7 @@ QVariant ImageListModelCondensed::data(const QModelIndex &index, int role) const
         QString condensed = "ID:%1 %2bpp %3x%4";
         condensed = condensed
                 .arg(img->nodeIndex())
-                .arg(img->getImageOriginalDepth())
+                .arg(static_cast<int>(img->getImageOriginalDepth()))
                 .arg(resolution.width())
                 .arg(resolution.height());
         res = condensed;
